# src/main/java/com/f5/taskf5/model/EditImage.java
package com.f5.taskf5.model;

public class EditImage {
    private String title;
    private String url;

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }
}